Second

+++ src/main/java/com/stackroute/service/MusicServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.service;
 
 import com.stackroute.domain.Music;
+import com.stackroute.exceptions.TrackAlreadyExistsException;
+import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.repository.MusicRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,9 +23,16 @@ public class MusicServiceImpl implements MusicService{
 
 
     @Override
-    public Music saveMusic(Music music) {
+    public Music saveMusic(Music music) throws TrackAlreadyExistsException {
+        if(musicRepository.existsById(music.getTrackId())){
+            throw new TrackAlreadyExistsException("User already exists");
+        }
 
         Music savedMusic= musicRepository.save(music);
+
+        if(music==null){
+            throw new TrackAlreadyExistsException("track doesnt exist");
+        }
         return savedMusic;
     }
 
@@ -33,25 +42,40 @@ public class MusicServiceImpl implements MusicService{
     }
 
     @Override
-    public Music updateComment(Music music,int trackId){
+    public Music updateComment(Music music,int trackId) throws TrackNotFoundException {
+        if(!musicRepository.existsById(trackId)){
+            throw new TrackNotFoundException("Track to update not found");
+        }
 
         Music updateMusic = musicRepository.save(music);
         return updateMusic;
     }
 
     @Override
-    public List<Music> deleteTrack( int trackId) {
-
+    public List<Music> deleteTrack( int trackId) throws TrackNotFoundException {
+        if(!musicRepository.existsById(trackId)){
+            throw new TrackNotFoundException("Track for deleting not found");
+        }
         musicRepository.deleteById(trackId);
         return musicRepository.findAll();
 
   }
 
     @Override
-    public Optional<Music> findById(int trackId) {
-
-           Optional<Music> music = musicRepository.findById(trackId);
+    public Optional<Music> findById(int trackId) throws TrackNotFoundException {
+        if(!musicRepository.existsById(trackId)){
+            throw new TrackNotFoundException("Track you searched not found ");
+        }
+        Optional<Music> music = musicRepository.findById(trackId);
         return music;
     }
+
+    @Override
+    public List<Music> findByName(String trackName) {
+
+        return musicRepository.findByName(trackName);
+
+    }
+
 
 }
